Remove duplicate trips, rotate the JWT key, extend decimal money to purchase
Duplicate trips
  Four groups of duplicate sale_details rows existed. Three were unambiguous and
  have been removed:

    - 489/490 (2025-04-22): 20 sale lines each, and all 20 pairs matched exactly
      on customer, birds, weight, rate, amount and payment - a double
      submission. Header 490 and its 20 lines removed.
    - 2114/2115 (2025-11-27): one identical line each. Header 2115 removed.
    - 1340/1341/1342 (2025-08-23): three headers with no sale lines at all,
      carrying phantom totals. Two removed, one kept.

  Removing them reversed 29,450 of double-counted pending across 21 customers,
  so each affected customer's balance_amount was reduced by their share before
  the ledger was cleared and rebuilt from the corrected data. Result: 56,120 ->
  56,099 sales, 3,695 -> 3,691 trips, balances 20,396,697.00 -> 20,367,247.00,
  and all 487 customers with ledger history reconcile to their stored balance
  with a worst gap of 0.00.

  Group 4/5 (2024-08-11) was deliberately NOT touched. Both rows are customer 28
  with 45 birds, fully paid, but one records 100 kg / 10,000 and the other
  92 kg / 9,200. That is not a duplicate but two different figures, most likely
  a correction where one should have replaced the other. Deleting the wrong one
  would lose real revenue, so it needs a decision. The UNIQUE constraint on
  (date, route, vehicle, driver) still cannot be added until it is resolved.

JWT key
  Rotated. The previous key was a literal in JWTUtils and is in git history, so
  it had to be treated as compromised. Verified by behaviour rather than by
  inspection: a token signed with the old key is now rejected with 401 and a
  fresh login succeeds. All existing sessions are invalidated by this.

Purchase, trading and supplier money (V4)
  V2 covered the sale -> ledger -> customer chain. These tables were worse:
  purchase.diesel, purchase.driver_expense, purchase.hamali and
  dc_detail.kilograms / rate / nos stored money and quantities as VARCHAR(255),
  with the rest split between INT and DOUBLE. Text cannot be summed, so this
  blocked the combined sale/purchase reporting.

  All are now DECIMAL. Checked first that every VARCHAR money column held only
  numeric values; purchase.hamali's four empty strings are set to NULL rather
  than being silently converted to zero.

  Verified against the restored data: Flyway applied V4, the application starts
  under ddl-auto=validate, no truncation warnings were raised, and purchase
  totals are unchanged at 1,744,540.00 - which dc_detail line amounts now sum to
  exactly, so purchase tallies as well.

  Adds indexes on sale (date+customer, date+route, date+driver), purchase
  (entry_date) and customer_ledger (customer+date) for the reporting queries.

  purchase_old and dc_detail_old are legacy and left untouched.

+++ src/main/java/com/app/dto/PurchaseDetailsDTO.java
@@ -1,11 +1,12 @@
 package com.app.dto;
 
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Data
 @AllArgsConstructor
 public class PurchaseDetailsDTO {
-	private Double totalAmount;
-    private Double paidAmount;
+	private BigDecimal totalAmount;
+    private BigDecimal paidAmount;
 }
